Fix silent failure when not logged in
When not logged in, or after the cookie expires, GOG.com will respond to
requests with a redirect to a login page; these were going unnoticed.
The target of the redirect isn't quite the same as LOGIN, so the code
uses the heuristic that a final URL ending in '/login' must mean this
happened.