Fix manifest parsing of removed games
GOG.com has removed a few games from its catalog that it once sold; the
known cases are all some sort of litigation or licensing issue. The
previous parser threw an exception when it encountered such an entry,
but now it just prints a warning.

Also, sneak in a refinement of the warning message for chunk XML with
invalid whole-file MD5s.

Reported-by: Luke Bartley